refactor: change controllers folder to routers
