+++ src/routers/medicationRoutes.ts
